Better new user experience. Fixes #81 - Also enabled precompiled header for Editor project

+++ Harmony/lib/src/netLink/src/smart_buffer.cc
@@ -96,7 +96,7 @@ void SmartBuffer::read(Socket* socket) {
 
         if(incomingBytes > freeBytes || !freeBytes) {
 
-            size_t newAllocSize = uMax((unsigned)(_allocSize * _reallocRatio), _usedSize + incomingBytes);
+            size_t newAllocSize = uMax((unsigned)(_allocSize * _reallocRatio), (unsigned int)(_usedSize + incomingBytes));
 
             void* newBuffer = realloc(_buffer, newAllocSize);
